Refactor contactController.js and userModel.js

+++ controllers/contactController.js
@@ -3,7 +3,7 @@ import contactModel from "../models/contactModel.js";
 
 
 const getContact = expressAsyncHandler(async (req, res) => {
-    const contact = await contactModel.find();
+    const contact = await contactModel.find({user_id:req.user.id});
     res.status(200).json(contact);
 });
 
@@ -18,7 +18,7 @@ const postContact =expressAsyncHandler(async (req,res,next)=>{
     }
 
     const contact = await contactModel.create({
-        name,email,phone
+        name,email,phone,user_id:req.user.id
     })
     res.status(201).json({contact})
 });
@@ -29,6 +29,10 @@ const getSpecific = expressAsyncHandler(async (req,res)=>{
         res.status(404);
         throw new Error("Contact Not found")
     }
+    if(contact.user_id.toString !== req.user.id){
+        res.status(403)
+        throw new Error("Not your contact")
+    }
     res.status(200).json(contact)
 });
 
@@ -37,6 +41,10 @@ const updateContact = expressAsyncHandler (async (req,res)=>{
     if(!contact){
         res.status(404);
         throw new Error("Contact Not found");
+    }
+    if(contact.user_id.toString !== req.user.id){
+        res.status(403)
+        throw new Error("Not your contact")
     }
     const newData = await contactModel.findByIdAndUpdate(req.params.id,req.body);
     res.status(200).json(newData);
@@ -48,6 +56,11 @@ const deleteContact = expressAsyncHandler(async (req,res)=>{
     if(!contact){
         res.status(404);
         throw new Error("Contact Not found");
+    }
+
+    if(contact.user_id.toString !== req.user.id){
+        res.status(403)
+        throw new Error("Not your contact")
     }
     await contactModel.deleteOne({_id:req.params.id});
     res.status(200).json(contact);
